requirements.txt
fastapi
uvicorn
onnxruntime
pywebview
numpy
pillow
python-multipart
jinja2
opencv-python
